Connectors: move module body out of index.ts into Connectors.ts
`index.ts` was carrying the entire module body — backend interface,
state machine, lifecycle internals, runtime + service builders, the
React bridge — while a sibling `Connectors.spec.ts` referenced
`./` indirectly. Rename the body to `Connectors.ts` and have `index.ts`
just re-export, mirroring the surrounding catalog convention. No code
changes — public API is preserved through the barrel.
